+++ src/types/config_files.ts
@@ -28,6 +28,7 @@ export interface CF_FKNODE_SETTINGS {
         enabled: boolean;
         freq: number;
     };
+    favoriteEditor: "vscode" | "sublime";
 }
 
 /**